+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_commands.py
@@ -562,6 +562,147 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
     @AKSCustomResourceGroupPreparer(
         random_name_length=17, name_prefix="clitest", location="westus2"
     )
+    def test_aks_create_with_addon_autoscaling(
+        self, resource_group, resource_group_location
+    ):
+        # reset the count so in replay mode the random names will start with 0
+        self.test_resources_count = 0
+        # kwargs for string formatting
+
+        aks_name = self.create_random_name("cliakstest", 16)
+        _, create_version = self._get_versions(resource_group_location)
+        self.kwargs.update(
+            {
+                "resource_group": resource_group,
+                "name": aks_name,
+                "ssh_key_value": self.generate_ssh_keys(),
+                "k8s_version": create_version,
+            }
+        )
+
+        # create
+        create_cmd = (
+            "aks create --resource-group={resource_group} --name={name} "
+            "--vm-set-type VirtualMachineScaleSets -c 1 "
+            "--enable-addon-autoscaling "
+            "--kubernetes-version={k8s_version} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/AKS-AddonAutoscalingPreview "
+            "--ssh-key-value={ssh_key_value} -o json"
+        )
+        self.cmd(
+            create_cmd,
+            checks=[
+                self.check("provisioningState", "Succeeded"),
+                self.check(
+                    "workloadAutoScalerProfile.verticalPodAutoscaler.enabled", True
+                ),
+                self.check(
+                    "workloadAutoScalerProfile.verticalPodAutoscaler.addonAutoscaling", "Enabled"
+                )
+            ],
+        )
+
+        # delete
+        cmd = (
+            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait"
+        )
+        self.cmd(
+            cmd,
+            checks=[
+                self.is_empty(),
+            ],
+        )
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(
+        random_name_length=17, name_prefix="clitest", location="westus2"
+    )
+    def test_aks_update_with_addon_autoscaling(
+        self, resource_group, resource_group_location
+    ):
+        # reset the count so in replay mode the random names will start with 0
+        self.test_resources_count = 0
+        # kwargs for string formatting
+
+        aks_name = self.create_random_name("cliakstest", 16)
+        _, create_version = self._get_versions(resource_group_location)
+        self.kwargs.update(
+            {
+                "resource_group": resource_group,
+                "name": aks_name,
+                "ssh_key_value": self.generate_ssh_keys(),
+                "k8s_version": create_version,
+            }
+        )
+
+        # create
+        create_cmd = (
+            "aks create --resource-group={resource_group} --name={name} "
+            "--vm-set-type VirtualMachineScaleSets -c 1 "
+            "--kubernetes-version={k8s_version} "
+            "--ssh-key-value={ssh_key_value} -o json"
+        )
+        self.cmd(
+            create_cmd,
+            checks=[
+                self.check("provisioningState", "Succeeded"),
+            ],
+        )
+
+        # update to enable
+        update_cmd = (
+            "aks update --resource-group={resource_group} --name={name} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/AKS-AddonAutoscalingPreview "
+            "--enable-addon-autoscaling -o json"
+        )
+        self.cmd(
+            update_cmd,
+            checks=[
+                self.check("provisioningState", "Succeeded"),
+                self.check(
+                    "workloadAutoScalerProfile.verticalPodAutoscaler.enabled", True
+                ),
+                self.check(
+                    "workloadAutoScalerProfile.verticalPodAutoscaler.addonAutoscaling", "Enabled"
+                ),
+            ],
+        )
+
+        # update to disable
+        update_cmd = (
+            "aks update --resource-group={resource_group} --name={name} "
+            "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/AKS-AddonAutoscalingPreview "
+            "--disable-addon-autoscaling -o json"
+        )
+        self.cmd(
+            update_cmd,
+            checks=[
+                self.check("provisioningState", "Succeeded"),
+                # disable addon autoscaling should not disable VPA
+                self.check(
+                    "workloadAutoScalerProfile.verticalPodAutoscaler.enabled", True
+                ),
+                self.check(
+                    "workloadAutoScalerProfile.verticalPodAutoscaler.addonAutoscaling", "Disabled"
+                ),
+            ],
+        )
+
+        # delete
+        cmd = (
+            "aks delete --resource-group={resource_group} --name={name} --yes --no-wait"
+        )
+        self.cmd(
+            cmd,
+            checks=[
+                self.is_empty(),
+            ],
+        )
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(
+        random_name_length=17, name_prefix="clitest", location="westus2"
+    )
     def test_aks_create_with_ingress_appgw_addon(
         self, resource_group, resource_group_location
     ):
@@ -3610,6 +3751,63 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         # delete the snapshot
         delete_snapshot_cmd = "aks snapshot delete --resource-group {resource_group} --name {snapshot_name} --yes --no-wait"
         self.cmd(delete_snapshot_cmd, checks=[self.is_empty()])
+
+    @AllowLargeResponse()
+    @AKSCustomResourceGroupPreparer(
+        random_name_length=17,
+        name_prefix="clitest",
+        location="eastus",
+        preserve_default_location=True,
+    )
+    def test_aks_skip_gpu_driver_install(self, resource_group, resource_group_location):
+        print(resource_group_location)
+        create_version, upgrade_version = self._get_versions(resource_group_location)
+        aks_name = self.create_random_name("cliakstest", 16)
+        nodepool_name = self.create_random_name("c", 6)
+
+        self.kwargs.update(
+            {
+                "resource_group": resource_group,
+                "name": aks_name,
+                "location": resource_group_location,
+                "nodepool_name": nodepool_name,
+                "k8s_version": upgrade_version,
+                "ssh_key_value": self.generate_ssh_keys(),
+                "windows_admin_username": "azureuser1",
+                "windows_admin_password": "replace-Password1234$",
+            }
+        )
+
+        # create an aks cluster
+        create_cmd = (
+            "aks create --resource-group {resource_group} --name {name} --location {location} "
+            "--node-count 2 "
+            "--windows-admin-username={windows_admin_username} --windows-admin-password={windows_admin_password} "
+            "--load-balancer-sku=standard --vm-set-type=virtualmachinescalesets --network-plugin=azure "
+            "-k {k8s_version} "
+            "--ssh-key-value={ssh_key_value} -o json"
+        )
+        self.cmd(
+            create_cmd, checks=[self.check("provisioningState", "Succeeded")]
+        )
+
+        # create nodepool from the cluster without gpu install
+        create_nodepool_cmd = (
+            "aks nodepool add --resource-group={resource_group} --cluster-name={name} --name={nodepool_name} --os-type windows --node-count 1 "
+            "--skip-gpu-driver-install "
+            "-k {k8s_version} -o json"
+        )
+        self.cmd(
+            create_nodepool_cmd,
+            checks=[self.check("provisioningState", "Succeeded"),
+                    self.check('gpuProfile.installGpuDriver', False)],
+        )
+
+        # delete the original AKS cluster
+        self.cmd(
+            "aks delete -g {resource_group} -n {name} --yes --no-wait",
+            checks=[self.is_empty()],
+        )
 
     # @AllowLargeResponse()
     # @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
@@ -10965,121 +11163,85 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
         assert len(role["rules"]) > 0
 
     @AllowLargeResponse()
-    @AKSCustomResourceGroupPreparer(
-        random_name_length=17, name_prefix="clitest", location="westus2"
-    )
-    def test_aks_trustedaccess_rolebinding(
-        self, resource_group, resource_group_location
-    ):
+    @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix='clitest', location='westus2')
+    def test_aks_create_with_trustedaccess_rolebinding(self, resource_group, resource_group_location):
         # reset the count so in replay mode the random names will start with 0
         self.test_resources_count = 0
         # kwargs for string formatting
-        aks_name = self.create_random_name("cliakstest", 16)
+        aks_name = self.create_random_name('cliakstest', 16)
 
-        self.kwargs.update(
-            {
-                "resource_group": resource_group,
-                "name": aks_name,
-                "location": resource_group_location,
-                "resource_type": "Microsoft.ContainerService/ManagedClusters",
-                "vm_size": "Standard_D4s_v3",
-                "node_count": 1,
-                "ssh_key_value": self.generate_ssh_keys(),
-            }
-        )
+        self.kwargs.update({
+            'resource_group': resource_group,
+            'name': aks_name,
+            'location': resource_group_location,
+            'resource_type': 'Microsoft.ContainerService/ManagedClusters',
+            'vm_size': 'Standard_D4s_v3',
+            'node_count': 1,
+            'ssh_key_value': self.generate_ssh_keys(),
+        })
 
-        create_cmd = " ".join(
-            [
-                "aks",
-                "create",
-                "--resource-group={resource_group}",
-                "--name={name}",
-                "--location={location}",
-                "--node-vm-size {vm_size}",
-                "--node-count {node_count}",
-                "--ssh-key-value={ssh_key_value}",
-                "--aks-custom-headers AKSHTTPCustomFeatures=Microsoft.ContainerService/TrustedAccessPreview",
-            ]
-        )
-        self.cmd(
-            create_cmd,
-            checks=[
-                self.check("provisioningState", "Succeeded"),
-            ],
-        )
+        create_cmd = ' '.join([
+            'aks', 'create', '--resource-group={resource_group}', '--name={name}', '--location={location}',
+            '--node-vm-size {vm_size}',
+            '--node-count {node_count}',
+            '--ssh-key-value={ssh_key_value}'
+        ])
 
-        binding_name = "testbinding"
-        node_rg_cmd = (
-            'aks list -g {resource_group} --query "[0].nodeResourceGroup" -o tsv'
-        )
-        node_rg = self.cmd(node_rg_cmd).output.strip()
-        self.kwargs.update({"node_rg": node_rg, "binding_name": binding_name})
+        self.cmd(create_cmd, checks=[
+            self.check('provisioningState', 'Succeeded'),
+        ])
 
-        vmss_cmd = 'vmss list -g {node_rg} --query "[0].id" -o tsv'
-        vmss_id = self.cmd(vmss_cmd).output.strip()
-        self.kwargs.update({"vmss_id": vmss_id})
+        list_binding_cmd = 'aks trustedaccess rolebinding list ' \
+            '--cluster-name={name} ' \
+            '--resource-group={resource_group}'
+        self.cmd(list_binding_cmd, checks=[
+            self.is_empty(),
+        ])
 
-        # test create rolebinding
-        create_ta_binding_cmd = " ".join(
-            [
-                "aks",
-                "trustedaccess",
-                "rolebinding",
-                "create",
-                "-g {resource_group}",
-                "--cluster-name {name}",
-                "-n {binding_name}",
-                "-s {vmss_id}",
-                "--roles Microsoft.Compute/virtualMachineScaleSets/test-node-reader,Microsoft.Compute/virtualMachineScaleSets/test-admin",
-            ]
-        )
-        binding = self.cmd(
-            create_ta_binding_cmd,
-            checks=[
-                self.check("name", binding_name),
-                self.check(
-                    "type",
-                    "Microsoft.ContainerService/managedClusters/trustedAccessRoleBindings",
-                ),
-                self.check("sourceResourceId", vmss_id),
-            ],
-        ).get_output_in_json()
-        assert len(binding["roles"]) == 2
-        time.sleep(20)  # wait for binding creation
+        subscription_id = self.get_subscription_id()
+        sid = '/subscriptions/{}/resourceGroups/{}/providers/Microsoft.MachineLearningServices/workspaces/fake_workspace'.format(subscription_id, resource_group)
+        self.kwargs.update({
+            'sid': sid
+        })
+        create_binding_cmd = 'aks trustedaccess rolebinding create ' \
+            '--resource-group={resource_group} ' \
+            '--cluster-name={name} ' \
+            '-n testbinding ' \
+            '-r {sid} ' \
+            '--roles Microsoft.MachineLearningServices/workspaces/mlworkload'
+        self.cmd(create_binding_cmd)
+        self.cmd(list_binding_cmd, checks=[
+            self.check('[0].type', 'Microsoft.ContainerService/managedClusters/trustedAccessRoleBindings'),
+            self.check('[0].name', 'testbinding'),
+        ])
 
-        # test list rolebinding
-        list_binding_cmd = "aks trustedaccess rolebinding list --cluster-name {name} -g {resource_group}"
-        listed_bindings = self.cmd(list_binding_cmd).get_output_in_json()
-        assert len(listed_bindings) == 1
+        get_binding_cmd = 'aks trustedaccess rolebinding show ' \
+            '-n testbinding ' \
+            '--resource-group={resource_group} ' \
+            '--cluster-name={name}'
+        self.cmd(get_binding_cmd, checks=[
+            self.check('type', 'Microsoft.ContainerService/managedClusters/trustedAccessRoleBindings'),
+            self.check('name', 'testbinding'),
+            self.check('provisioningState', 'Succeeded'),
+        ])
 
-        # test get rolebinding
-        get_binding_cmd = "aks trustedaccess rolebinding show --cluster-name {name} -g {resource_group} -n {binding_name}"
-        got_binding = self.cmd(get_binding_cmd).get_output_in_json()
-        assert got_binding["name"] == binding_name
-        assert len(got_binding["roles"]) == 2
-        assert got_binding["sourceResourceId"] == vmss_id
-        assert (
-            got_binding["type"]
-            == "Microsoft.ContainerService/managedClusters/trustedAccessRoleBindings"
-        )
+        update_binding_cmd = 'aks trustedaccess rolebinding update ' \
+            '--cluster-name={name} ' \
+            '-n testbinding ' \
+            '--resource-group={resource_group} ' \
+            '--roles Microsoft.MachineLearningServices/workspaces/mlworkload'
+        self.cmd(update_binding_cmd, checks=[
+            self.check('roles[0]', 'Microsoft.MachineLearningServices/workspaces/mlworkload'),
+        ])
 
-        # test update rolebinding
-        update_binding_cmd = "aks trustedaccess rolebinding update -g {resource_group} --cluster-name {name} -n {binding_name} \
-            --roles Microsoft.Compute/virtualMachineScaleSets/test-node-reader"
-        updated_binding = self.cmd(update_binding_cmd).get_output_in_json()
-        assert updated_binding["name"] == binding_name
-        assert len(updated_binding["roles"]) == 1
-        assert (
-            updated_binding["roles"][0]
-            == "Microsoft.Compute/virtualMachineScaleSets/test-node-reader"
-        )
-
-        # test delete rolebinding
-        delete_binding_cmd = "aks trustedaccess rolebinding delete -g {resource_group} --cluster-name {name} -n {binding_name} -y"
+        delete_binding_cmd = 'aks trustedaccess rolebinding delete ' \
+            '--cluster-name={name} ' \
+            '-n testbinding ' \
+            '--resource-group={resource_group} -y'
         self.cmd(delete_binding_cmd)
-        time.sleep(20)  # wait for binding deleting
-        listed_bindings = self.cmd(list_binding_cmd).get_output_in_json()
-        assert len(listed_bindings) == 0
+        self.cmd(list_binding_cmd, checks=[
+            self.is_empty(),
+        ])
 
     @AllowLargeResponse()
     @AKSCustomResourceGroupPreparer(